pyproject.toml
[project]
name = "pitch-deck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.4",
    "langchain[google-genai]>=1.2.10",
    "langgraph>=1.0.8",
    "python-dotenv>=1.2.1",
]
